More flexibility to engage pages metadata checks
This allows more flexibility to add additional metadata checks.
Passing `self.additional_metadata_validation` will allow the project to
add custom metadata validation without changing things in discourse
module.

It was conceived as an "additional" validation, because we don't want to
write all the checks all over again, only the ones which are optional,
e.g. `language` metadata key, which is only required in u.com, not on
cn.u.com or jp.u.com where takeovers are always in the same language. So
we don't need to send such sentry errors.